% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ble.R
\name{ble}
\alias{ble}
\title{Generate a synthetic data from a stub precursor object}
\usage{
ble(stub, rows, method, ctrl = list(), ...)
}
\arguments{
\item{stub}{the stub object which synthetic data is generated.}

\item{rows}{the number of elements to generate.}

\item{method}{the method to use in the generation of synthetic data. Must be
one of \code{"agnostic"} (default) or \code{"empirical"}.}

\item{ctrl}{a named list of control parameters for generating the synthetic
data. See \code{\link{stubble_ctrl}}.}

\item{...}{named individual control parameters, which take precedence over
those in the \code{ctrl} list.}
}
\value{
An object of the same class used in \code{\link{stub}}.
}
\description{
\code{ble()} is an internal function of \code{\link{stubble}} that is used to generate
synthetic data from an precursor object that has been generated by \link{stub}.
}
\details{
\code{ble()} calls an internal S3 generic function, \code{ble_()}, with built-in
methods for the \code{list} and \code{data.frame} base R data types, in addition to
\code{data.table} and \code{tbl_df} data types, where the required \emph{data.table} and
\emph{tibble} packages are installed.
}
\examples{
ble(stub(penguins_ext))
ble(stub(penguins_ext), rows = 10)

}
\seealso{
\link{stubble_ctrl}
}
\concept{ecdf}
\concept{empirical}
\concept{resample}
\concept{simulate}
\concept{simulated}
\concept{simulation}
\keyword{datagen}
