% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/col-gen.R
\name{col_gen}
\alias{col_gen}
\title{Generate simple synthetic data from an R vector}
\usage{
col_gen(col, nrows, ...)
}
\arguments{
\item{col}{the vector from which the class of the synthetic data is taken.}

\item{nrows}{the number of elements to generate.}

\item{...}{control parameters for ranges and valid levels/characters in the
synthetic data. See [stubble::control].}
}
\value{
Returns a vector of the same class as `col` with `nrows` elements.
}
\description{
`col_gen()` does the real work for [stubblise()]. Currently supported column
classes are `numeric`, `integer`, `character`, `factor`, `logical`,
`POSIXct` and `Date`. There is limited support for `list` (returns a list
with all `NA`s).
}
\examples{
\dontrun{
library(stubble)
}
col_gen(iris$Sepal.length)
col_gen(iris$Species)

}
