% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emperor.R
\name{emperor}
\alias{emperor}
\title{Generate synthetic data using the empirical distribution of an R vector}
\usage{
emperor(col, elements = length(col), index = 1L, control = list(), ...)
}
\arguments{
\item{col}{the vector from which the type of the synthetic data is taken.}

\item{elements}{the number of elements to generate.}

\item{index}{the index of the column (i.e. the position in the data frame)
for the purposes of extracting control parameters.}

\item{control}{a named list of control parameters for generating the
synthetic data. See \link{control}.}

\item{...}{named individual control parameters, which take precedence over
those in the \code{control} list.}
}
\value{
Returns a vector of the same class as \code{col} with \code{elements} elements.
}
\description{
\code{emperor()} does the real work for \code{\link[=stubblise]{stubblise()}}. Currently supported column
classes are \code{double}, \code{integer}, \code{character}, \code{factor}, \code{ordered}, \code{logical},
\code{POSIXct}, \code{POSIXlt}, \code{Date} and \code{IDate}. There is limited support for \code{list}
(returns a list with all \code{NA}s).
}
\details{
\code{emperor()} calls an internal S3 generic function, \code{emperor_()},
with built-in methods for base R vector types.

All methods return values sampled from the cumulative empirical distribution
function.
}
\note{
This function is entirely experimental and should not be used yet. There are
currently risks of data leakage and many of the methods do not yet work.
}
\examples{
emperor(iris$Sepal.length)
emperor(iris$Species)

}
\seealso{
\code{\link[=emperor_control]{emperor_control()}}
}
\concept{ecdf}
\concept{empirical}
\concept{resample}
\concept{simulate}
\concept{simulated}
\concept{simulation}
\keyword{datagen}
\keyword{distribution}
\keyword{manip}
