% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-col.R
\name{gen_col}
\alias{gen_col}
\title{Generate simple synthetic data from an R vector}
\usage{
gen_col(col, elements = 10L, index = 1L, ...)
}
\arguments{
\item{col}{the vector from which the type of the synthetic data is taken.}

\item{elements}{the number of elements to generate.}

\item{index}{the index of the column for the purposes of extracting control
parameters.}

\item{...}{control parameters for ranges and valid levels/characters in the
synthetic data. See \link{control}.}
}
\value{
Returns a vector of the same class as \code{col} with \code{elements} elements.
}
\description{
\code{gen_col()} does the real work for \code{\link[=stubblise]{stubblise()}}. Currently supported column
classes are \code{numeric}, \code{integer}, \code{character}, \code{factor}, \code{logical},
\code{POSIXct} and \code{Date}. There is limited support for \code{list} (returns a list
with all \code{NA}s).
}
\examples{
gen_col(iris$Sepal.length)
gen_col(iris$Species)

}
