% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-col-control.R
\name{gen_col_control}
\alias{gen_col_control}
\alias{control}
\title{Control parameters for gen_col()}
\usage{
gen_col_control(
  unique = FALSE,
  int_min = 0L,
  int_max = 100L,
  dbl_min = 0,
  dbl_max = 100,
  dbl_rng_kind = "Wichmann-Hill",
  dbl_round = NA,
  dbl_signif = NA,
  chr_min = 0L,
  chr_max = 10L,
  chr_sym = list(c(letters, LETTERS, as.character(0:9),
    unlist(strsplit("!\\"#$\%&'()*+, -./:;<=>?@[]^_`{|}~", "")))),
  chr_force_unique = FALSE,
  chr_force_unique_attempts = 10L,
  fct_lvls = list(letters[1:4]),
  fct_use_lvls = NULL,
  lgl_vals = list(c(TRUE, FALSE)),
  date_origin = "1970-01-01",
  date_max = Sys.Date(),
  dttm_max = Sys.time(),
  dttm_tz = "UTC",
  def_class = "numeric",
  old_ctrl = as.list(NULL),
  index = NA,
  ...
)
}
\arguments{
\item{unique}{Single logical value or logical vector indicating whether synthetic values should be unique within the column. When a vector, the relevant element is chosen by the \code{index} argument to \link{gen_col}.}

\item{int_min}{Minmium values for integer generation.}

\item{int_max}{Maximum values for integer generation.}

\item{dbl_min}{Minimum values for real/numeric/double generation}

\item{dbl_max}{Maximum values for real/numeric/double generation}

\item{dbl_rng_kind}{The random number generation algorithm to use for real/numeric/double values. Takes values allowed by \code{\link[=RNGkind]{RNGkind()}}. Defaults to \code{"Wichmann-Hill"}, which is slower than other generators but is much less likely to produce duplicate values.}

\item{dbl_round}{Number of decimal places to round to. \code{\link[=round]{round()}} and then \code{\link[=signif]{signif()}} are applied in sequence (see \code{dbl_signif}). If \code{NA} (default), no rounding is applied.}

\item{dbl_signif}{Number of significant digits to round to. \code{\link[=round]{round()}} and then \code{\link[=signif]{signif()}} are applied in sequence (see \code{dbl_round}). If \code{NA} (default), no rounding is applied.}

\item{chr_min}{The minimum number of characters in a generated string.}

\item{chr_max}{The maximum number of characters in a generated string.}

\item{chr_sym}{A character vector of allowed symbols for generated strings.}

\item{chr_force_unique}{Logical value indicating whether, after a failure to generate a unique synthetic character vector, the algorithm should attempt to regenerate the vector. If \code{TRUE}, there will be \code{chr_force_unique_attempts} attempts.}

\item{chr_force_unique_attempts}{Number of attempts to make to generate a unique synthetic character vector.}

\item{fct_lvls}{Levels attribute for synthetic factors. Should always be a list with character vectors as elements.}

\item{fct_use_lvls}{Allowed levels for synthetic factor data. Should always be a list with character vectors as elements. Each element of \code{fct_use_lvls} should be a subset of the corresponding element of \code{fct_lvls}.}

\item{lgl_vals}{Allowed values of generated logicals.}

\item{date_origin}{The reference date for generated dates and times.}

\item{date_max}{Max value for generated dates.}

\item{dttm_max}{Max value for generated date-times.}

\item{dttm_tz}{Timezone for generated date-times. Defaults to \code{"UTC"}, but \code{Sys.timezone()} may be more appropriate for some users.}

\item{def_class}{A default vector class to return when all else fails. Currently not used.}

\item{old_ctrl}{A set of control parameters to inherit unless explicitly overwritten in the current call.}

\item{index}{Default \code{NA}. If not \code{NA}, the function will return list in which elements apply to a single column (i.e. elements are not necessarily lists). Mostly for internal use to handle passing control parameters between \code{gen_col_} S3 methods.}

\item{...}{Further control parameters permitting extension of the \code{gen_col_} S3 methods.}
}
\description{
This help file describes the available parameters to control the output of
\code{\link[=stubblise]{stubblise()}}.
}
\details{
Generation of synthetic data in stubble is very simple. Numbers,
dates and times are sampled uniformly within a range, while strings and
factors are constructed from lists of symbols or levels sampled with equal
probability.

Various control parameters allow some user influence over the sets of
allowed values. Available parameters are listed and described above. In
addition to \code{gen_col_control()}, these parameters are exposed directly via
\code{\link[=stubblise]{stubblise()}} and \code{\link[=gen_col]{gen_col()}}, and may be passed as arguments with the names
given above to either function.

Control parameters may be common to all or may be specified per-column.
Arguments passed to \code{gen_col_control()} or directly via \code{stubblise()} or
\code{gen_col()} are interpreted according to type:
\itemize{
\item \emph{Single-element vectors} mean that the same parameter value applies
to each column;
\item \emph{Multi-element vectors} mean that each element is matched (with
recycling) as the parameter for the corresponding column;
}
To pass a vector parameter for a column, such as a vector of allowed factor
levels, wrap the vector in a list, e.g. \code{list(letters[1:4])}. (Internally,
all arguments passed to \code{gen_col_control()}, whether directly or indirectly,
are sanitised with \code{as.list()}.)
}
