% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/penguins_ext.R
\docType{data}
\name{penguins_ext}
\alias{penguins_ext}
\title{Data on adult penguins near Palmer Station, Antarctica}
\format{
An object of class \code{data.frame} with 344 rows and 10 columns.
}
\source{
{All data were derived from the \href{https://allisonhorst.github.io/palmerpenguins/}{palmerpenguins} package, under that package's \href{https://creativecommons.org/publicdomain/zero/1.0/}{CC0-1.0 license}. Original sources are listed below.}

{Adélie penguins: Palmer Station Antarctica LTER and K. Gorman. 2020. Structural size measurements and isotopic signatures of foraging among adult male and female Adélie penguins (Pygoscelis adeliae) nesting along the Palmer Archipelago near Palmer Station, 2007-2009 ver 5. Environmental Data Initiative} \url{https://doi.org/10.6073/pasta/98b16d7d563f265cb52372c8ca99e60f}

{Gentoo penguins: Palmer Station Antarctica LTER and K. Gorman. 2020. Structural size measurements and isotopic signatures of foraging among adult male and female Gentoo penguin (Pygoscelis papua) nesting along the Palmer Archipelago near Palmer Station, 2007-2009 ver 5. Environmental Data Initiative} \url{https://doi.org/10.6073/pasta/7fca67fb28d56ee2ffa3d9370ebda689}

{Chinstrap penguins: Palmer Station Antarctica LTER and K. Gorman. 2020. Structural size measurements and isotopic signatures of foraging among adult male and female Chinstrap penguin (Pygoscelis antarcticus) nesting along the Palmer Archipelago near Palmer Station, 2007-2009 ver 6. Environmental Data Initiative} \url{https://doi.org/10.6073/pasta/c14dfcfada8ea13a17536e73eb6fbe9e}

{Originally published in: Gorman KB, Williams TD, Fraser WR (2014) Ecological Sexual Dimorphism and Environmental Variability within a Community of Antarctic Penguins (Genus Pygoscelis). PLoS ONE 9(3): e90081. doi:10.1371/journal.pone.0090081}
}
\usage{
penguins_ext
}
\description{
Measurements of physical characteristics and nesting success of penguins of
several species on islands in the Palmer Archipelago. Derived from the
\href{https://allisonhorst.github.io/palmerpenguins/}{palmerpenguins} package.
}
\details{
These data are an \strong{ext}ended version of the \link[palmerpenguins]{penguins}
data, using addition variables from \link[palmerpenguins]{penguins_raw}.
Variables were chosen to represent several data types (character, factor,
numeric, integer, logical, and Date) for use in examples for stubble (see e.g. \code{\link[=stub]{stub()}}).

The data include the following variables:
\describe{
\item{\code{id}}{a character string denoting the unique ID for each individual in dataset}
\item{\code{species}}{a factor denoting penguin species (Adélie, Chinstrap and Gentoo)}
\item{\code{island}}{a factor denoting island in Palmer Archipelago, Antarctica (Biscoe, Dream or Torgersen)}
\item{\code{bill_length_mm}}{a numeric denoting bill length (millimeters)}
\item{\code{bill_depth_mm}}{a numeric denoting bill depth (millimeters)}
\item{\code{flipper_length_mm}}{an integer denoting flipper length (millimeters)}
\item{\code{body_mass_g}}{an integer denoting body mass (grams)}
\item{\code{sex}}{a factor denoting penguin sex (female, male)}
\item{\code{clutch_completion}}{logical denoting whether the nest was observed with a full clutch (2 eggs)}
\item{\code{date_egg}}{a Date denoting the date study nest observed with 1 egg (sampled)}
}
(Variable descriptions were adapted from the palmerpenguins package.)
}
\examples{
summary(penguins_ext)
}
\keyword{datasets}
