% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stub.R
\name{stub}
\alias{stub}
\title{Generate a stub object from which synthetic data may be generated}
\usage{
stub(x, rows = lengths(x), method = "agnostic", ctrl = list(), ...)
}
\arguments{
\item{x}{the vector from which the type of the synthetic data is taken.}

\item{rows}{the number of elements to generate.}

\item{ctrl}{a named list of control parameters for generating the synthetic
data. See \link{ctrl}.}

\item{...}{named individual control parameters, which take precedence over
those in the \code{ctrl} list.}
}
\value{
A list containing three elements:
\itemize{
\item \code{ctrl} Any non-default control parameters set.
\item \code{dtype} A zero-dimension object of the same class as the source data.
\item \code{vars} Attributes relating to each element/column of the source data.
}
}
\description{
\code{stub()} is an internal function of \code{\link{stubble}} that is used to generate a
list-type object, which in turn, may be used by the \code{\link{ble}} internal function
to generate synthetic data. Calling \code{stub()} directly allows the user to
generate an object containing all the necessary information to generate a
synthetic version of the source data, without actually doing so. This may be
useful when sharing synthetic versions of your source data with others.
}
\details{
\code{stub()} calls an internal S3 generic function, \code{stub_()}, with built-in
methods for the \code{list} and \code{data.frame} base R data types, in addition to
\code{data.table} and \code{tbl_df} data types, where the required \emph{data.table} and
\emph{tibble} packages are installed.
}
\note{
Custom row names from objects with the class attribute \code{data.frame} are not
currently captured by \code{stub()} and will not be reproduced in any data
generated from the resulting \code{stub} object.
}
\examples{
stub(iris)
stub(iris, rows = 10)

}
\seealso{
\link{stubble_ctrl}
}
\concept{ecdf}
\concept{empirical}
\concept{resample}
\concept{simulate}
\concept{simulated}
\concept{simulation}
\keyword{datagen}
