% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stubble.R
\docType{package}
\name{stubble-package}
\alias{stubble}
\alias{stubble-package}
\title{stubble: Create simple synthetic data from a data frame-like objects}
\description{
stubble takes a data frame-like object as input and returns an object of the same class containing randomly generated (synthetic/simulated) data consistent with the original column types.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/bjcairns/stubble}
  \item Report bugs at \url{https://github.com/bjcairns/stubble/issues}
}

}
\author{
\strong{Maintainer}: Benjamin J. Cairns \email{ben.cairns@ndph.ox.ac.uk} (\href{https://orcid.org/0000-0001-7994-8213}{ORCID})

Authors:
\itemize{
  \item Benjamin G. Feakins \email{benjamin.feakins@ndph.ox.ac.uk} (\href{https://orcid.org/0000-0002-3928-6750}{ORCID})
}

Other contributors:
\itemize{
  \item University of Oxford [copyright holder]
}

}
\keyword{datagen}
\keyword{internal}
\keyword{manip}
