% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emperor-control.R
\name{emperor_control}
\alias{emperor_control}
\alias{control}
\title{Control parameters for emperor()}
\usage{
emperor_control(
  p_na = NA_real_,
  emp_sw = 0.5,
  breaks = "FD",
  tail_exc = 0.025,
  fuzz_ecdf = TRUE,
  fuzz_sca = 0.1,
  fuzz_ht = TRUE,
  n_exc = 10,
  p_exc = 0.01,
  drop_lev = TRUE,
  dttm_tz = "UTC",
  old_ctrl = as.list(NULL),
  index = NA_integer_,
  ...
)
}
\arguments{
\item{p_na}{Proportion of values set to \code{NA}; defaults to \code{NA}.}

\item{emp_sw}{Value determining whether spline or resampling methods are used
in the the generation of synthetic data. When the unique fraction of a column
is above this value spline-based methods will be used. Conversely, when it is
below this value resampling methods will be used. Hence, setting it to 1 will
ensure that resampling methods will always be used, while setting it to 0
will ensure that spline-based methods will always be used. Defaults to 50\%
(\code{0.5}). will always be used. When set to 0 ECDF-based methods will always be
used.}

\item{breaks}{Number of breaks to use for spline-based interpolation of the
empirical cumulative distribution function. Can be either \code{"FD"} (default) or
a whole number. When set to \code{"FD"} the number of breaks is calculated using
the Freedman-Diaconis method.}

\item{tail_exc}{Quantile tail size to be omitted from sampling at each end
of the empirical cumulative distribution function. Defaults to 5\% (\code{0.05})
at each end (tail) of the distribution.}

\item{fuzz_ecdf}{Should the values sampled from the ECDF be 'fuzzed'
through the addition of random normal noise? Defaults to \code{TRUE}.}

\item{fuzz_sca}{Scaling factor for the random noise added to the ECDF data.
Defaults to \code{0.1}, i.e. one 10th the range of the original data.
of the original data.}

\item{fuzz_ht}{Should all fuzzed parameters be kept within the bounds set
by \code{tail_exc}? Defaults to \code{TRUE}.}

\item{n_exc}{Observation prevalence below which values will be excluded from
simulations. Defaults to 10.}

\item{p_exc}{Observation prevalence below which values will be excluded from
simulations. Defaults to 1\% (\code{0.01})}

\item{drop_lev}{Parameter indicating whether empty factor levels should be
dropped from simlated factors and ordered factors. Defauls to \code{TRUE}}

\item{dttm_tz}{Timezone for generated date-times. Defaults to \code{"UTC"}, but
\code{\link[=Sys.timezone]{Sys.timezone()}} may be more appropriate for some users.}

\item{old_ctrl}{A set of control parameters to inherit unless explicitly
overwritten in the current call.}

\item{index}{Default \code{NA}. If not \code{NA}, the function will return list in
which elements apply to a single column (i.e. elements are not necessarily
lists). Mostly for internal use to handle passing control parameters between
\code{emperor_} S3 methods.}

\item{...}{Further control parameters permitting extension of the \code{emperor_}
S3 methods.}
}
\description{
This help file describes the available parameters to control the output of
\code{\link[=stubblise]{stubblise()}}. The function which assembles these parameters into a list,
\code{emperor_control()}, is accessible to users but often will not need to be
called directly.
}
\details{
Generation of synthetic data in stubble is very simple. Numbers,
dates and times are sampled uniformly within a range, while strings and
factors are constructed from lists of symbols or levels sampled with equal
probability.

Various control parameters allow some user influence over the sets of
allowed values. Available parameters are listed and described above. In
addition to \code{emperor_control()}, these parameters are exposed directly via
\code{\link[=stubblise]{stubblise()}} and \code{\link[=emperor]{emperor()}}, and may be passed as arguments with the names
given above to either function.

Control parameters may be common to all or may be specified per-column.
Arguments passed to \code{emperor_control()} or directly via \code{stubblise()} or
\code{emperor()} are interpreted according to type:
\itemize{
\item \emph{Single-element vectors} mean that the same parameter value applies
to each column;
\item \emph{Multi-element vectors} mean that each element is matched (with
recycling) as the parameter for the corresponding column;
}
To pass a vector parameter for a column, such as a vector of allowed factor
levels, wrap the vector in a list, e.g. \code{list(letters[1:4])}. (Internally,
all arguments passed to \code{emperor_control()}, whether directly or indirectly,
are sanitised with \code{as.list()}.)
}
\seealso{
\code{\link[=emperor]{emperor()}}
}
