% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emperor-control.R
\name{emperor_control}
\alias{emperor_control}
\alias{control}
\title{Control parameters for emperor()}
\usage{
emperor_control(
  p_na = NA,
  tailsize = 0.05,
  dbl_round = NA,
  dbl_signif = NA,
  dttm_tz = "UTC",
  old_ctrl = as.list(NULL),
  index = NA,
  ...
)
}
\arguments{
\item{p_na}{Proportion of values set to \code{NA}; defaults to \code{NA}.}

\item{tailsize}{Quantile size to be omitted from sampling at each end of
the empirical cumulative distribution function.}

\item{dbl_round}{Number of decimal places to round to. \code{\link[=round]{round()}} and then
\code{\link[=signif]{signif()}} are applied in sequence (see \code{dbl_signif}, below). If \code{NA}
(default), no rounding is applied.}

\item{dbl_signif}{Number of significant digits to round to. \code{\link[=round]{round()}} and
then \code{\link[=signif]{signif()}} are applied in sequence (see \code{dbl_round}). If \code{NA}
(default), no rounding is applied.}

\item{dttm_tz}{Timezone for generated date-times. Defaults to \code{"UTC"}, but
\code{\link[=Sys.timezone]{Sys.timezone()}} may be more appropriate for some users.}

\item{old_ctrl}{A set of control parameters to inherit unless explicitly
overwritten in the current call.}

\item{index}{Default \code{NA}. If not \code{NA}, the function will return list in
which elements apply to a single column (i.e. elements are not necessarily
lists). Mostly for internal use to handle passing control parameters between
\code{emperor_} S3 methods.}

\item{...}{Further control parameters permitting extension of the \code{emperor_}
S3 methods.}
}
\description{
This help file describes the available parameters to control the output of
\code{\link[=stubblise]{stubblise()}}. The function which assembles these parameters into a list,
\code{emperor_control()}, is accessible to users but often will not need to be
called directly.
}
\details{
Generation of synthetic data in stubble is very simple. Numbers,
dates and times are sampled uniformly within a range, while strings and
factors are constructed from lists of symbols or levels sampled with equal
probability.

Various control parameters allow some user influence over the sets of
allowed values. Available parameters are listed and described above. In
addition to \code{emperor_control()}, these parameters are exposed directly via
\code{\link[=stubblise]{stubblise()}} and \code{\link[=emperor]{emperor()}}, and may be passed as arguments with the names
given above to either function.

Control parameters may be common to all or may be specified per-column.
Arguments passed to \code{emperor_control()} or directly via \code{stubblise()} or
\code{emperor()} are interpreted according to type:
\itemize{
\item \emph{Single-element vectors} mean that the same parameter value applies
to each column;
\item \emph{Multi-element vectors} mean that each element is matched (with
recycling) as the parameter for the corresponding column;
}
To pass a vector parameter for a column, such as a vector of allowed factor
levels, wrap the vector in a list, e.g. \code{list(letters[1:4])}. (Internally,
all arguments passed to \code{emperor_control()}, whether directly or indirectly,
are sanitised with \code{as.list()}.)
}
\seealso{
\code{\link[=emperor]{emperor()}}
}
\author{
Benjamin G. Feakins, \email{benjamin.feakins@ndph.ox.ac.uk}
}
