% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-col.R
\name{gen_col}
\alias{gen_col}
\title{Generate simple synthetic data from an R vector}
\usage{
gen_col(col, nrows, ...)
}
\arguments{
\item{col}{the vector from which the class of the synthetic data is taken.}

\item{nrows}{the number of elements to generate.}

\item{...}{control parameters for ranges and valid levels/characters in the
synthetic data. See [stubble::control].}
}
\value{
Returns a vector of the same class as `col` with `nrows` elements.
}
\description{
`gen_col()` does the real work for [stubblise()]. Currently supported column
classes are `numeric`, `integer`, `character`, `factor`, `logical`,
`POSIXct` and `Date`. There is limited support for `list` (returns a list
with all `NA`s).
}
\examples{
\dontrun{
library(stubble)
}
gen_col(iris$Sepal.length)
gen_col(iris$Species)

}
