% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stubble-ctrl.R
\name{stubble_ctrl}
\alias{stubble_ctrl}
\title{Control parameters for stubble}
\usage{
stubble_ctrl(
  p_na = NA_real_,
  emp_sw = 0.1,
  tail_exc = 0.025,
  fuzz_spl = TRUE,
  fuzz_spl_sca = 0.05,
  n_exc = 10,
  p_exc = 0.05,
  fuzz_samp = TRUE,
  drop_lev = TRUE,
  dttm_tz = "UTC",
  old_ctrl = list(),
  index = NA_integer_,
  ...
)
}
\arguments{
\item{p_na}{Proportion of values set to \code{NA}; defaults to \code{NA}, meaning that
the proportions present in any generated data will roughly match those of the
source data.}

\item{emp_sw}{Value determining whether spline or resampling methods are used
in the the generation of synthetic data. When the unique fraction of a column
is above this value spline-based methods will be used. Conversely, when it is
below this value resampling methods will be used. Hence, setting it to 1 will
ensure that resampling methods will always be used, while setting it to 0
will ensure that spline-based methods will always be used. Defaults to 50\%
(\code{0.5}). will always be used. When set to 0 ECDF-based methods will always be
used.}

\item{tail_exc}{Quantile tail size to be omitted from sampling at each end
of the empirical cumulative distribution function. Defaults to 2.5\% (\code{0.025})
at each end (tail) of the distribution.}

\item{fuzz_spl}{Should the values sampled from the ECDF by \code{\link{ble_spline}}
be 'fuzzed' through the addition of random normal noise? Defaults to \code{TRUE}.}

\item{fuzz_spl_sca}{The scaling factor for the standard deviation of the
random noise applied when \code{fuzz_spl} is set to \code{TRUE}. Defaults to \code{0.05},
i.e. 5\% of the standard deviation of the source data.}

\item{n_exc}{Observation prevalence below which values will be excluded from
simulations. Defaults to 10.}

\item{p_exc}{Observation prevalence below which values will be excluded from
simulations. Defaults to 1\% (\code{0.01})}

\item{fuzz_samp}{Should the probability weights sampled from the distribution
of values by \code{\link{stub_sample}} be 'fuzzed' through the addition of random
normal noise? Defaults to \code{TRUE}.}

\item{drop_lev}{Parameter indicating whether empty factor levels should be
dropped from simlated factors and ordered factors. Defauls to \code{TRUE}}

\item{dttm_tz}{Timezone for generated date-times. Defaults to \code{"UTC"}, but
\code{\link[=Sys.timezone]{Sys.timezone()}} may be more appropriate for some users.}

\item{old_ctrl}{A set of control parameters to inherit unless explicitly
overwritten in the current call.}

\item{index}{Default \code{NA}. If not \code{NA}, the function will return list in
which elements apply to a single column (i.e. elements are not necessarily
lists). Mostly for internal use to handle passing control parameters between
\code{emperor_} S3 methods.}

\item{...}{Further control parameters permitting extension of the \code{stubble}
S3 methods.}
}
\description{
This help file describes the available parameters to control the output of
\code{\link[=stubble]{stubble()}}. The function is accessible to users but often will not need to
be called directly.
}
\seealso{
\code{\link[=stubble]{stubble()}}
}
