% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stubble-ctrl.R
\name{stubble_ctrl}
\alias{stubble_ctrl}
\title{Control parameters for stubble}
\usage{
stubble_ctrl(
  rng_kind = "Wichmann-Hill",
  p_na = NA_real_,
  dttm_tz = "UTC",
  agn_unique = FALSE,
  agn_int_min = 0L,
  agn_int_max = 100L,
  agn_int_list = NA_integer_,
  agn_int64_min = 0L,
  agn_int64_max = 100L,
  agn_dbl_min = 0,
  agn_dbl_max = 100,
  agn_dbl_round = NA_integer_,
  agn_dbl_signif = NA_integer_,
  agn_chr_min = 0L,
  agn_chr_max = 10L,
  agn_chr_sym = list(c(letters, LETTERS, 0:9,
    strsplit("!\\"#$\%&'()*+, -./:;<=>?@[]^_`{|}~", "")[[1]])),
  agn_chr_sep = "",
  agn_chr_try_unique = FALSE,
  agn_chr_try_unique_attempts = 10L,
  agn_chr_duplicated_nmax = NA_integer_,
  agn_fct_lvls = list(letters[1:4]),
  agn_fct_use_lvls = list(),
  agn_fct_force_unique = FALSE,
  agn_lgl_force_unique = FALSE,
  agn_date_origin = "1970-01-01",
  agn_date_min = agn_date_origin,
  agn_date_max = Sys.Date(),
  agn_dttm_min = agn_date_origin,
  agn_dttm_max = format(Sys.time(), tz = dttm_tz),
  agn_time_min = "00:00:00",
  agn_time_max = "23:59:59",
  emp_sw = 0.1,
  emp_tail_exc = 0.025,
  emp_fuzz_spl = 0.05,
  emp_n_exc = 10L,
  emp_p_exc = 0.05,
  emp_fuzz_samp = 0.05,
  emp_drop_lev = TRUE,
  old_ctrl = list(),
  assert_class = TRUE,
  index = NA_integer_,
  ...
)
}
\arguments{
\item{rng_kind}{The random number generation algorithm to use. Takes values
allowed by \code{\link[=RNGkind]{RNGkind()}}. Defaults to \code{"Wichmann-Hill"}, which is slower than
other generators but is much less likely to produce duplicate values.}

\item{p_na}{Proportion of values set to \code{NA}; defaults to \code{NA}, meaning that
the proportions present in any generated data will roughly match those of the
source data.}

\item{dttm_tz}{Timezone for generated date-times (\code{POSIXt} class variables).
Defaults to \code{"UTC"}, but \code{\link[=Sys.timezone]{Sys.timezone()}} may be more appropriate for some
users.}

\item{agn_unique}{Single logical value or logical vector indicating whether
synthetic values should be unique within the column. When a vector, the
relevant element is chosen by the \code{index} argument to \code{ble_agnostic}.}

\item{agn_int_min}{Minmium values for integer generation.}

\item{agn_int_max}{Maximum values for integer generation.}

\item{agn_int_list}{An integer vector of allowed values for integer
generation. If \code{NA} (the default), this is ignored. If non-\code{NA},}

\item{agn_int64_min}{Minimum values for integer64 generation (requires
\emph{bit64} package).}

\item{agn_int64_max}{Maximum values for integer64 generation (requires
\emph{bit64} package).
\code{agn_int_list} overrides \code{agn_int_min}/\code{agn_int_max}.}

\item{agn_dbl_min}{Minimum values for real/numeric/double generation.}

\item{agn_dbl_max}{Maximum values for real/numeric/double generation.}

\item{agn_dbl_round}{Number of decimal places to round to. \code{\link[=round]{round()}} and then
\code{\link[=signif]{signif()}} are applied in sequence (see \code{agn_dbl_signif}, below). If \code{NA}
(default), no rounding is applied.}

\item{agn_dbl_signif}{Number of significant digits to round to. \code{\link[=round]{round()}} and
then \code{\link[=signif]{signif()}} are applied in sequence (see \code{agn_dbl_round}). If \code{NA}
(default), no rounding is applied.}

\item{agn_chr_min}{The minimum number of characters in a generated string.}

\item{agn_chr_max}{The maximum number of characters in a generated string.}

\item{agn_chr_sym}{A character vector of allowed symbols for generated
strings.}

\item{agn_chr_sep}{A separator for symbols in generated strings; defaults to
\code{""} (an empty string).}

\item{agn_chr_try_unique}{Logical value indicating whether, after a failure
to generate a unique synthetic character vector, the algorithm should attempt
to regenerate duplicates. If \code{TRUE}, there will be
\code{agn_chr_try_unique_attempts} attempts.}

\item{agn_chr_try_unique_attempts}{Number of attempts to make to generate a
unique synthetic character vector.}

\item{agn_chr_duplicated_nmax}{Value (greater than one) for the \code{nmax}
parameter of \code{\link[=duplicated]{duplicated()}} when enforcing uniqueness. Defaults to \code{NA}, the
default for \code{duplicated()}.}

\item{agn_fct_lvls}{Levels attribute for synthetic factors. Should always be
a list with character vectors as elements.}

\item{agn_fct_use_lvls}{Allowed levels for synthetic factor data. Should
always be a list with character vectors as elements. Each element of
\code{agn_fct_use_lvls} should be a subset of the corresponding element of
\code{agn_fct_lvls}.}

\item{agn_fct_force_unique}{Force \code{gen_col()} to attempt to return a factor
vector with unique elements. Will return an error unless
\code{length(agn_fct_use_lvls)} is greater than or equal to the \code{elements}
argument of \code{gen_col()} (equivalently, the \code{nrows} argument of
\code{stubblise()}).}

\item{agn_lgl_force_unique}{Force \code{gen_col()} to attempt to return a logical
vector with unique elements. Will return an error unless
\code{length(agn_lgl_lvls)} (usually equal to 2) is greater than or equal to the
\code{elements} argument of \code{gen_col()} (equivalently, the \code{nrows} argument of
\code{stubblise()}).}

\item{agn_date_origin}{The reference date for generated dates and times (\code{Date},
\code{IDate} and \code{POSIXt} class variables).}

\item{agn_date_min}{Minimum value for generated dates (\code{Date} and \code{IDate}
class variables).}

\item{agn_date_max}{Maximum value for generated dates (\code{Date} and \code{IDate}
class variables).}

\item{agn_dttm_min}{Minimum value for generated date-times (\code{POSIXt} class
variables).}

\item{agn_dttm_max}{Maximum value for generated date-times (\code{POSIXt} class
variables).}

\item{agn_time_min}{Minimum value for generated times (\code{ITime} class
variables).}

\item{agn_time_max}{Maximum value for generated times (\code{ITime} class
variables).}

\item{emp_sw}{Value determining whether spline or sampling methods are used
in the the generation of synthetic data. When the unique fraction of a column
is above this value spline-based methods will be used. Conversely, when it is
below this value sampling methods will be used. Hence, setting it to \code{1} will
ensure that sampling methods will always be used, while setting it to \code{0}
will ensure that spline-based methods will always be used. Defaults to 10\%
(\code{0.1}).}

\item{emp_tail_exc}{Quantile tail size to be omitted from sampling at each
end of the empirical cumulative distribution function. Defaults to 2.5\%
(\code{0.025}).}

\item{emp_fuzz_spl}{Scaling factor for the standard deviation of the
random noise applied to continuous variables prior to sampling the empirical
cumulative distribution function by \code{stub_spline}. Defaults to \code{0.05}, i.e.
5\% of the standard deviation of the source data. Setting this value to \code{0}
switches off obfuscation of the values evaluated by \code{stub_spline}.}

\item{emp_n_exc}{Observation prevalence below which values will be excluded
by \code{stub_sample}. Defaults to \code{10}.}

\item{emp_p_exc}{Observation prevalence below which values will be excluded
by \code{stub_sample}. Defaults to 5\% (\code{0.05}).}

\item{emp_fuzz_samp}{Scaling factor for the obfuscation of the sample
proportions measured by \code{stub_sample}. Defaults to \code{0.05}. Larger values add
more noise. Setting this value to \code{0} switches off obfuscation of the sample
proportions recorded by \code{stub_sample}.}

\item{emp_drop_lev}{Parameter indicating whether empty factor levels should
be dropped from simulated factors and ordered factors. Defaults to \code{TRUE}.}

\item{old_ctrl}{A set of control parameters to inherit unless explicitly
overwritten in the current call.}

\item{assert_class}{Logical value indicating whether control parameter
classes should be checked. Mostly intended for internal use. Defaults to
\code{TRUE}.}

\item{index}{Default \code{NA}. If not \code{NA}, the function will return list in
which elements apply to a single column, i.e. elements are not necessarily
lists. Mostly for internal use to handle passing control parameters between
S3 methods.}

\item{...}{Further control parameters permitting extension of the \code{stubble}
S3 methods.}
}
\description{
This help file describes the available parameters to control the output of
\code{\link[=stubble]{stubble()}}. The function is accessible to users but often will not need to
be called directly.
}
\seealso{
\code{\link[=stubble]{stubble()}}
}
