% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stubblise.R
\name{stubblise}
\alias{stubblise}
\alias{stubblize}
\title{Create simple simulated data from a tibble}
\usage{
stubblise(tbbl, nrows = 10, ...)

stubblize(tbbl, nrows = 10, ...)
}
\arguments{
\item{tbbl}{the tibble to emulate. Can have 0 rows.}

\item{nrows}{the number of simulated rows to generate.}

\item{...}{additional arguments to control the data format. See
[stubblise_control].}
}
\value{
A tibble, with the same structure as `tbbl`, but containing random
data. Any columns which are not recognised as being of a standard vector
class will have `NA` for all rows and the column type will be `logical`,
not the original vector class.
}
\description{
`stubblise()` (equivalently, `stubblize()`) generates synthetic (i.e.
simulated) data to match the structure of a given tibble. See also
[col_gen()], which does most of the work, and [control] for information on
user-modifiable parameters.
}
\details{
One intened use of `stubblise()` is in generating simple test data
for R package development. The function can create very simple synthetic
data from most tibbles where columns have standard R vector types. Currently
list-columns are not fully supported, and this and any other unsupported
types will return a column of NAs.

Data are randomly generated by sampling uniformly from a numeric or integer
range or from a selection of e.g. characters or factor levels. The
parameters controlling this sampling can be modified by the user. At
present, the same parameters apply to all columns, and there is no attempt
to make the synthetic data look anything like the original data in `tbbl`.

The function `stubblize()` is a synonym for `stubblise()` for those who
find the "z" more natural or convenient.
}
\examples{

\dontrun{
library(stubble)
}
stubblise(iris)
stubblise(iris, fct_lvls = levels(iris$Species))

}
