% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stubble.R, R/stubblise.R
\name{stubble}
\alias{stubble}
\alias{stubblise}
\alias{stubblize}
\title{Create simple synthetic data from a data frame-like objects

\code{stubblise()} (equivalently, \code{stubblize()}) generates synthetic (i.e.
simulated) data to match the structure of a given data frame-like object
(including data.frames, tibbles, data.tables, and lists). See also
\link{stubble_ctrl} for information on user-modifiable parameters.}
\usage{
stubblise(x, rows = lengths(x), method = "agnostic", ctrl = list(), ...)

stubblize(x, rows = lengths(x), method = "agnostic", ctrl = list(), ...)
}
\arguments{
\item{x}{the data frame-like object to emulate. Can have 0 rows.}

\item{rows}{the number of rows to generate.}

\item{method}{the method to use in the generation of synthetic data. Must be
one of \code{"agnostic"} (default) or \code{"empirical"}.}

\item{ctrl}{a named list of control parameters for generating the
synthetic data. See \link{control}.}

\item{...}{named individual control parameters, which take precedence over
those in the \code{control} list.}
}
\value{
A data frame-like object with the same class as \code{x} (unless \code{x} is a
\code{list} or otherwise is coerceable to a \code{data.frame}, in which case the
result is a \code{data.frame}). Any columns which are not recognised as being of
a standard vector type will have \code{NA} for all rows and the column type will
be \code{numeric}, i.e. not the original vector type.
}
\description{
Create simple synthetic data from a data frame-like objects

\code{stubblise()} (equivalently, \code{stubblize()}) generates synthetic (i.e.
simulated) data to match the structure of a given data frame-like object
(including data.frames, tibbles, data.tables, and lists). See also
\link{stubble_ctrl} for information on user-modifiable parameters.
}
\details{
One intended use of \code{stubblise()} is in generating test data for R
package development. The function can create very simple synthetic data from
most data frame-like objects where columns have base R vector types. Lists
of vectors of unequal lengths (i.e. not coerceable to \code{data.frame}) are
allowed. Currently list-columns are not fully supported, and this and any
other unsupported types will return a column of \code{NA}s (for unsupported
types, the resulting column of \code{NA}s will have type \code{numeric}).

Data are randomly generated by sampling uniformly from a numeric or integer
range or from a selection of e.g. characters or factor levels. The
parameters controlling this sampling can be modified by the user. Control
parameters can be lists, in which case elements are matched (with recycling)
to corresponding columns.

Note that although \code{x} can have 0 rows, there is no benefit to passing
\code{stubblise()} an object with 0 rows. The computation time is not dependent on
the number of rows in \code{x} (thanks to the magic of S3 methods and lazy
evaluation).  Computation time does increase, approximately linearly, with
the number of columns in \code{x} and with the number of rows required in the
result (\code{rows}).

The function \code{stubblize()} is a synonym for \code{stubblise()}.
}
\section{Sensitive data}{


By default, outputs from \code{stubblise()} should contain no sensitive data,
\strong{unless the column names and types are themselves sensitive}. Attributes of
\code{x} are not preserved by \code{stubblise()} by default, and any attributes that
persist will have been regenerated by coercion (e.g. by \code{as.data.frame()})
and are therefore unlikely to be sensitive information.

It is possible to leak sensitive data via a "stubblised" data frame if the
values of the \link{control} parameters are themselves derived from sensitive
data. It is recommended to leave these parameters at their default values in
any circumstance that involves sensitive data (e.g. personal data of any
sort), or otherwise to ensure that they are set to non-informative and/or
non-sensitive values.
}

\examples{
# A simple examples producing nonsense data
syn_penguins_ext_0 <- stubblise(penguins_ext)

# Example with species factor levels
syn_penguins_ext_1 <- stubblise(
  penguins_ext,
  agn_fct_lvls = list(levels(penguins_ext$species))
)

# Example producing data more comparable to the real penguins_ext
# Uses dplyr
mins <- as.numeric(dplyr::summarise_if(penguins_ext, is.numeric, min))
maxs <- as.numeric(dplyr::summarise_if(penguins_ext, is.numeric, max))
lvls <- levels(penguins_ext$species)
rows <- dplyr::count(penguins_ext, species)$n
syn_penguins_ext_2 <- purrr::map2_dfr(
  rows, lvls,
  ~ stubblise(
    penguins_ext, rows = .x,
    control = list(
      agn_dbl_min = mins, agn_dbl_max = maxs, agn_dbl_round = 1L,
      agn_fct_lvls = list(lvls), agn_fct_use_lvls = list(.y)
    )
  )
)
}
